Changed punch frame data, added kick attack, started animation controller fixes, and added city exterior scene with terrain
Made punch faster with less lag and added kick attack. Some of the identified issues with the animator were fixed, but others were discovered in testing(and will be fixed in next push). The terrain asset is in the city exterior scene, though I'm currently dissatisfied with the appearance of the terrain and am planning to clean it up later to get the desired look. Expect a visual update with texture painting, some of the messy areas of the terrain cleaned up, lake, and trees.