Cleaned up terrain and imported the Free Rocks Pack
Smoothed the terrain and flattened the top of the hill. Painted textures for grass and mountains. Added trees, but they are currently invisible from far away while  the game is running and completely invisible while in scene view. I still don't know what is causing this or how to change it since I'm still learning how to use all of the terrain features