Create PlayerControllerCollectionFinal prefab
The final prefab uses the model with the new animations, though these animations are not yet implemented in the animator that controls them. The public references in the inventory were replaced, and the errors that previously popped up are fixed. The player controller can now be dragged or dropped into any scene without having to plug in these public references. Also included are physics changes to make the player faster. I'm pushing this now since it will likely take me much longer to get the transitions for the animator sorted out. The prefab will just pull from the animator when I push again, without needing to make another player controller prefab.