Camera Adjustments and Added NPC models with animations to project
Adjusted camera voidance strategies. It still has some issues, but I think I've struck the best balance that can be achieved with the time remaining between avoiding walls and avoiding other smaller obstacles too quickly. Added the models for the NPCs to the project files, but the new versions still need to be added into the scene.