Set up ladders and Ledge in Tony's Lair, Fixed Player Getting Stuck Soft lock
The ladders and ledge in Tony's Lair are now fully functional and ready to go. I also added some colliders for the ramps and fryers for the platforming, but the rest of Tony's Lair still needs to be given hitboxes. I found a bug where the player could get stuck on objects they jumped onto that were too small for the Grounded Check raycasts to hit, and they would get stuck unable to move, soft locking the game. I added code to the TPC script to prevent this by forcing the player state to be grounded after 3 seconds, and another bug was born, allowing the player to walk on the air when jumping down from sufficiently high places. I worked around this by adding another condition to check if the player's squared velocity was less than 0.1.